Refactored the live activity control panel out of <LiveActivity>
