+++ src/components/live_activity.css
@@ -1,10 +1,3 @@
-.liveController {
-    padding: 20px 30px;
-    width: 95%;
-    justify-content: space-around;
-    font-size: 30px;
-    vertical-align: middle;
-}
 
 .stepCard {
     margin-top: 0px;
@@ -14,32 +7,8 @@
     display: none;
 }
 
-.timerDivider {
-    margin: 10px !important;
-}
-
-.timerContainer{
-    font-size: 55px;
-    text-align: center;
-}
-
 .stepCard {
     margin-top: 10px;
-}
-
-.nextStepButton {
-    margin-top: 20px;
-}
-
-.currentStepButton {
-    /* defined inline */
-    font-weight: 400 !important;
-    font-size: 22px !important;
-}
-
-.tooltip {
-    font-size: 14px !important;
-    display: block !important;
 }
 
 @media (min-width:640px) {
@@ -47,25 +16,7 @@
         display: block;
     }
 
-    .liveController {
-        width: 640px;
-    }
-
-    .timer {
-        margin: 30px !important;
-    }
-
     .stepCard {
         margin-top: 120px;
-    }
-
-    .pageControls {
-        display: flex;
-        align-items: center;
-        justify-content: space-around;
-    }
-
-    .nextStepButton {
-        margin-top: 0;
     }
 }
